fix: config-ui blueprints service pack 2.0 (#2630)
* fix: enhance stage task name text details

* fix: cleanup mock objects from blueprint detail

* chore: apply linting cleanups

* feat: enable pipeline logfile download

* fix: resolve lint dependencies with bp validate

* fix: terminate regex match on repo name validation
