+++ config-ui/package.json
@@ -24,6 +24,7 @@
     "dayjs": "^1.10.7",
     "dotenv": "^10.0.0",
     "dotenv-webpack": "^7.0.3",
+    "file-saver": "^2.0.5",
     "jetbrains-mono": "^1.0.6",
     "react": "17.0.2",
     "react-dom": "17.0.2",
