+++ config-ui/src/pages/blueprints/blueprint-detail.jsx
@@ -21,6 +21,7 @@ import { useParams, useHistory } from 'react-router-dom'
 import { DEVLAKE_ENDPOINT } from '@/utils/config'
 import request from '@/utils/request'
 import dayjs from '@/utils/time'
+import { saveAs } from 'file-saver'
 import {
   Button,
   Elevation,
@@ -41,412 +42,18 @@ import {
 import { NullBlueprint } from '@/data/NullBlueprint'
 import { NullPipelineRun } from '@/data/NullPipelineRun'
 import { Providers, ProviderLabels, ProviderIcons } from '@/data/Providers'
-
-// import {
-//   WorkflowSteps,
-//   WorkflowAdvancedSteps,
-//   DEFAULT_DATA_ENTITIES,
-//   DEFAULT_BOARDS,
-// } from '@/data/BlueprintWorkflow'
+import { StageStatus, TaskStatus, TaskStatusLabels, StatusColors, StatusBgColors } from '@/data/Task'
 
 import Nav from '@/components/Nav'
 import Sidebar from '@/components/Sidebar'
 import Content from '@/components/Content'
-import TaskActivity from '@/components/pipelines/TaskActivity'
+// import TaskActivity from '@/components/pipelines/TaskActivity'
 import CodeInspector from '@/components/pipelines/CodeInspector'
 import StageLane from '@/components/pipelines/StageLane'
+import { ToastNotification } from '@/components/Toast'
 
 import useBlueprintManager from '@/hooks/useBlueprintManager'
 import usePipelineManager from '@/hooks/usePipelineManager'
-// import useConnectionManager from '@/hooks/useConnectionManager'
-// import { DataEntityTypes } from '@/data/DataEntities'
-
-const StageStatus = {
-  PENDING: 'Pending',
-  COMPLETE: 'Complete',
-  FAILED: 'Failed',
-  ACTIVE: 'In Progress',
-}
-
-const TaskStatus = {
-  COMPLETE: 'TASK_COMPLETED',
-  FAILED: 'TASK_FAILED',
-  ACTIVE: 'TASK_RUNNING',
-  RUNNING: 'TASK_RUNNING',
-  CREATED: 'TASK_CREATED',
-  PENDING: 'TASK_CREATED',
-  CANCELLED: 'TASK_CANCELLED',
-}
-
-const TaskStatusLabels = {
-  [TaskStatus.COMPLETE]: 'Succeeded',
-  [TaskStatus.FAILED]: 'Failed',
-  [TaskStatus.ACTIVE]: 'In Progress',
-  [TaskStatus.RUNNING]: 'In Progress',
-  [TaskStatus.CREATED]: 'Created (Pending)',
-  [TaskStatus.PENDING]: 'Created (Pending)',
-  [TaskStatus.CANCELLED]: 'Cancelled',
-}
-
-const StatusColors = {
-  PENDING: '#292B3F',
-  COMPLETE: '#4DB764',
-  FAILED: '#E34040',
-  ACTIVE: '#7497F7',
-}
-
-// eslint-disable-next-line no-unused-vars
-const StatusBgColors = {
-  PENDING: 'transparent',
-  COMPLETE: '#EDFBF0',
-  FAILED: '#FEEFEF',
-  ACTIVE: '#F0F4FE',
-}
-
-// eslint-disable-next-line no-unused-vars
-const TEST_BLUEPRINT = {
-  ...NullBlueprint,
-  id: 1,
-  name: 'DevLake Daily Blueprint',
-  createdAt: new Date().toLocaleString(),
-  updatedAt: new Date().toLocaleString(),
-}
-
-// eslint-disable-next-line no-unused-vars
-const TEST_CONNECTIONS = [
-  {
-    id: 0,
-    provider: Providers.GITHUB,
-    name: 'Merico GitHub',
-    dataScope: 'merico-dev/ake, merico-dev/lake-website',
-    dataEntities: ['code', 'ticket', 'user'],
-  },
-  {
-    id: 0,
-    provider: Providers.JIRA,
-    name: 'Merico JIRA',
-    dataScope: 'Sprint Dev Board, DevLake Sync Board ',
-    dataEntities: ['ticket'],
-  },
-]
-
-const TEST_RUN = {
-  id: null,
-  status: TaskStatus.RUNNING,
-  statusLabel: TaskStatusLabels[TaskStatus.RUNNING],
-  icon: <Spinner size={18} intent={Intent.PRIMARY} />,
-  startedAt: '7/7/2022, 5:31:33 PM',
-  duration: '1 min',
-  stage: 'Stage 1',
-  tasksTotal: 5,
-  tasksFinished: 8,
-  // totalTasks: 13,
-  error: null,
-}
-
-// eslint-disable-next-line no-unused-vars
-const EMPTY_RUN = {
-  id: null,
-  status: TaskStatus.CREATED,
-  statusLabel: TaskStatusLabels[TaskStatus.RUNNING],
-  icon: null,
-  startedAt: Date.now(),
-  duration: '0 min',
-  stage: 'Stage 1',
-  tasksTotal: 0,
-  tasksFinished: 0,
-  // totalTasks: 0,
-  error: null,
-}
-
-// eslint-disable-next-line no-unused-vars
-const TEST_BLUEPRINT_API_RESPONSE = {
-  name: 'DEVLAKE (Hourly)',
-  mode: 'NORMAL',
-  plan: [
-    [
-      {
-        plugin: 'github',
-        subtasks: [
-          'collectApiRepo',
-          'extractApiRepo',
-          'collectApiIssues',
-          'extractApiIssues',
-          'collectApiPullRequests',
-          'extractApiPullRequests',
-          'collectApiComments',
-          'extractApiComments',
-          'collectApiEvents',
-          'extractApiEvents',
-          'collectApiPullRequestCommits',
-          'extractApiPullRequestCommits',
-          'collectApiPullRequestReviews',
-          'extractApiPullRequestReviewers',
-          'collectApiCommits',
-          'extractApiCommits',
-          'collectApiCommitStats',
-          'extractApiCommitStats',
-          'enrichPullRequestIssues',
-          'convertRepo',
-          'convertIssues',
-          'convertCommits',
-          'convertIssueLabels',
-          'convertPullRequestCommits',
-          'convertPullRequests',
-          'convertPullRequestLabels',
-          'convertPullRequestIssues',
-          'convertIssueComments',
-          'convertPullRequestComments',
-        ],
-        options: {
-          connectionId: 1,
-          owner: 'e2corporation',
-          repo: 'incubator-devlake',
-          transformationRules: {
-            issueComponent: '',
-            issuePriority: '',
-            issueSeverity: '',
-            issueTypeBug: '',
-            issueTypeIncident: '',
-            issueTypeRequirement: '',
-            prComponent: '',
-            prType: '',
-          },
-        },
-      },
-      {
-        plugin: 'gitextractor',
-        subtasks: null,
-        options: {
-          repoId: 'github:GithubRepo:1:506830252',
-          url: 'https://git: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@github.com/e2corporation/incubator-devlake.git',
-        },
-      },
-    ],
-    [
-      {
-        plugin: 'refdiff',
-        subtasks: null,
-        options: {
-          tagsLimit: 10,
-          tagsOrder: '',
-          tagsPattern: '',
-        },
-      },
-    ],
-  ],
-  enable: true,
-  cronConfig: '0 0 * * *',
-  isManual: false,
-  settings: {
-    version: '1.0.0',
-    connections: [
-      {
-        connectionId: 1,
-        plugin: 'github',
-        scope: [
-          {
-            entities: ['CODE', 'TICKET'],
-            options: {
-              owner: 'e2corporation',
-              repo: 'incubator-devlake',
-            },
-            transformation: {
-              prType: '',
-              prComponent: '',
-              issueSeverity: '',
-              issueComponent: '',
-              issuePriority: '',
-              issueTypeRequirement: '',
-              issueTypeBug: '',
-              issueTypeIncident: '',
-              refdiff: {
-                tagsOrder: '',
-                tagsPattern: '',
-                tagsLimit: 10,
-              },
-            },
-          },
-        ],
-      },
-    ],
-  },
-  id: 1,
-  createdAt: '2022-07-11T10:23:38.908-04:00',
-  updatedAt: '2022-07-11T10:23:38.908-04:00',
-}
-
-const TEST_STAGES = [
-  {
-    id: 1,
-    name: 'stage-1',
-    title: 'Stage 1',
-    status: StageStatus.COMPLETED,
-    icon: <Icon icon='tick-circle' size={14} color={StatusColors.COMPLETE} />,
-    tasks: [
-      {
-        id: 0,
-        provider: 'jira',
-        icon: ProviderIcons[Providers.JIRA](14, 14),
-        title: 'JIRA',
-        caption: 'STREAM Board',
-        duration: '4 min',
-        subTasksCompleted: 25,
-        recordsFinished: 1234,
-        message: 'All 25 subtasks completed',
-        status: TaskStatus.COMPLETE,
-      },
-      {
-        id: 0,
-        provider: 'jira',
-        icon: ProviderIcons[Providers.JIRA](14, 14),
-        title: 'JIRA',
-        caption: 'LAKE Board',
-        duration: '4 min',
-        subTasksCompleted: 25,
-        recordsFinished: 1234,
-        message: 'All 25 subtasks completed',
-        status: TaskStatus.COMPLETE,
-      },
-    ],
-    stageHeaderClassName: 'complete',
-  },
-  {
-    id: 2,
-    name: 'stage-2',
-    title: 'Stage 2',
-    status: StageStatus.PENDING,
-    icon: <Spinner size={14} intent={Intent.PRIMARY} />,
-    tasks: [
-      {
-        id: 0,
-        provider: 'jira',
-        icon: ProviderIcons[Providers.JIRA](14, 14),
-        title: 'JIRA',
-        caption: 'EE Board',
-        duration: '5 min',
-        subTasksCompleted: 25,
-        recordsFinished: 1234,
-        message: 'Subtask 5/25: Extracting Issues',
-        status: TaskStatus.ACTIVE,
-      },
-      {
-        id: 0,
-        provider: 'jira',
-        icon: ProviderIcons[Providers.JIRA](14, 14),
-        title: 'JIRA',
-        caption: 'EE Bugs Board',
-        duration: '0 min',
-        subTasksCompleted: 0,
-        recordsFinished: 0,
-        message: 'Invalid Board ID',
-        status: TaskStatus.FAILED,
-      },
-    ],
-    stageHeaderClassName: 'active',
-  },
-  {
-    id: 3,
-    name: 'stage-3',
-    title: 'Stage 3',
-    status: StageStatus.PENDING,
-    icon: null,
-    tasks: [
-      {
-        id: 0,
-        provider: 'github',
-        icon: ProviderIcons[Providers.GITHUB](14, 14),
-        title: 'GITHUB',
-        caption: 'merico-dev/lake',
-        duration: null,
-        subTasksCompleted: 0,
-        recordsFinished: 0,
-        message: 'Subtasks pending',
-        status: TaskStatus.CREATED,
-      },
-    ],
-    stageHeaderClassName: 'pending',
-  },
-  {
-    id: 4,
-    name: 'stage-4',
-    title: 'Stage 4',
-    status: StageStatus.PENDING,
-    icon: null,
-    tasks: [
-      {
-        id: 0,
-        providr: 'github',
-        icon: ProviderIcons[Providers.GITHUB](14, 14),
-        title: 'GITHUB',
-        caption: 'merico-dev/lake',
-        duration: null,
-        subTasksCompleted: 0,
-        recordsFinished: 0,
-        message: 'Subtasks pending',
-        status: TaskStatus.CREATED,
-      },
-    ],
-    stageHeaderClassName: 'pending',
-  },
-]
-
-const TEST_HISTORICAL_RUNS = [
-  {
-    id: 0,
-    status: 'TASK_COMPLETED',
-    statusLabel: 'Completed',
-    statusIcon: <Icon icon='tick-circle' size={14} color={Colors.GREEN5} />,
-    startedAt: '05/25/2022 0:00 AM',
-    completedAt: '05/25/2022 0:15 AM',
-    duration: '15 min',
-  },
-  {
-    id: 1,
-    status: 'TASK_COMPLETED',
-    statusLabel: 'Completed',
-    statusIcon: <Icon icon='tick-circle' size={14} color={Colors.GREEN5} />,
-    startedAt: '05/25/2022 0:00 AM',
-    completedAt: '05/25/2022 0:15 AM',
-    duration: '15 min',
-  },
-  {
-    id: 2,
-    status: 'TASK_FAILED',
-    statusLabel: 'Failed',
-    statusIcon: <Icon icon='delete' size={14} color={Colors.RED5} />,
-    startedAt: '05/25/2022 0:00 AM',
-    completedAt: '05/25/2022 0:00 AM',
-    duration: '0 min',
-  },
-  {
-    id: 3,
-    status: 'TASK_COMPLETED',
-    statusLabel: 'Completed',
-    statusIcon: <Icon icon='tick-circle' size={14} color={Colors.GREEN5} />,
-    startedAt: '05/25/2022 0:00 AM',
-    completedAt: '05/25/2022 0:15 AM',
-    duration: '15 min',
-  },
-  {
-    id: 4,
-    status: 'TASK_COMPLETED',
-    statusLabel: 'Completed',
-    statusIcon: <Icon icon='tick-circle' size={14} color={Colors.GREEN5} />,
-    startedAt: '05/25/2022 0:00 AM',
-    completedAt: '05/25/2022 0:15 AM',
-    duration: '15 min',
-  },
-  {
-    id: 5,
-    status: 'TASK_FAILED',
-    statusLabel: 'Failed',
-    statusIcon: <Icon icon='delete' size={14} color={Colors.RED5} />,
-    startedAt: '05/25/2022 0:00 AM',
-    completedAt: '05/25/2022 0:00 AM',
-    duration: '0 min',
-  },
-]
 
 const BlueprintDetail = (props) => {
   // eslint-disable-next-line no-unused-vars
@@ -454,8 +61,7 @@ const BlueprintDetail = (props) => {
   const { bId } = useParams()
 
   const [blueprintId, setBlueprintId] = useState()
-  // @todo: replace with live $blueprint from Hook
-  const [activeBlueprint, setActiveBlueprint] = useState(TEST_RUN)
+  const [activeBlueprint, setActiveBlueprint] = useState(NullBlueprint)
   // eslint-disable-next-line no-unused-vars
   const [blueprintConnections, setBlueprintConnections] = useState([])
   const [blueprintPipelines, setBlueprintPipelines] = useState([])
@@ -465,13 +71,14 @@ const BlueprintDetail = (props) => {
   const [showCurrentRunTasks, setShowCurrentRunTasks] = useState(true)
   const [showInspector, setShowInspector] = useState(false)
   const [currentStages, setCurrentStages] = useState([])
-  const [historicalRuns, setHistoricalRuns] = useState(TEST_HISTORICAL_RUNS)
+  const [historicalRuns, setHistoricalRuns] = useState([])
 
   const pollTimer = 5000
   const pollInterval = useRef()
   const [autoRefresh, setAutoRefresh] = useState(false)
 
   const [expandRun, setExpandRun] = useState(null)
+  const [isDownloading, setIsDownloading] = useState(false)
 
   const {
     // eslint-disable-next-line no-unused-vars
@@ -523,6 +130,8 @@ const BlueprintDetail = (props) => {
     allowedProviders,
     // eslint-disable-next-line no-unused-vars
     detectPipelineProviders,
+    logfile: pipelineLogFilename,
+    getPipelineLogfile
   } = usePipelineManager()
 
   const buildPipelineStages = useCallback((tasks = []) => {
@@ -597,11 +206,39 @@ const BlueprintDetail = (props) => {
         icon = <Icon icon='delete' size={14} color={Colors.RED5} />
         break
       case TaskStatus.CANCELLED:
+        icon = <Icon icon='undo' size={14} color={Colors.RED5} />
+        break
       case TaskStatus.CREATED:
+        icon = <Icon icon='stopwatch' size={14} color={Colors.GRAY3} />
         break
     }
     return icon
   }
+
+  const downloadPipelineLog = useCallback((pipeline) => {
+    console.log(`>>> DOWNLOADING PIPELINE #${pipeline?.id}  LOG...`, getPipelineLogfile(pipeline?.id))
+    setIsDownloading(true)
+    ToastNotification.clear()
+    let downloadStatus = 404
+    const checkDownloadStatus = async (pipeline) => {
+      const d = await request.get(getPipelineLogfile(pipeline?.id))
+      downloadStatus = d?.status
+    }
+    checkDownloadStatus()
+    if (pipeline?.id && downloadStatus === 200) {
+      saveAs(
+        getPipelineLogfile(pipeline?.id),
+        pipelineLogFilename
+      )
+      setIsDownloading(false)
+    } else if (pipeline?.id && downloadStatus === 404) {
+      ToastNotification.show({ message: 'Logfile not available', intent: 'danger', icon: 'error' })
+      setIsDownloading(false)
+    } else {
+      ToastNotification.show({ message: 'Pipeline Invalid or Missing', intent: 'danger', icon: 'error' })
+      setIsDownloading(false)
+    }
+  }, [getPipelineLogfile, pipelineLogFilename])
 
   useEffect(() => {
     setBlueprintId(bId)
@@ -610,7 +247,6 @@ const BlueprintDetail = (props) => {
 
   useEffect(() => {
     if (blueprintId) {
-      // @todo: enable blueprint data fetch
       fetchBlueprint(blueprintId)
       fetchAllPipelines()
     }
@@ -648,7 +284,6 @@ const BlueprintDetail = (props) => {
 
   useEffect(() => {
     console.log('>>>> FETCHED ALL PIPELINES..', pipelines, activeBlueprint?.id)
-    //  {id: 5, status: 'TASK_FAILED', statusLabel: 'Failed', statusIcon: <Icon icon='delete' size={14} color={Colors.RED5} />,startedAt: '05/25/2022 0:00 AM', completedAt: '05/25/2022 0:00 AM', duration: '0 min' },
     setBlueprintPipelines(
       pipelines.filter((p) => p.blueprintId === activeBlueprint?.id)
     )
@@ -657,7 +292,6 @@ const BlueprintDetail = (props) => {
   useEffect(() => {
     console.log('>>>> RELATED BLUEPRINT PIPELINES..', blueprintPipelines)
     setLastPipeline(blueprintPipelines[0])
-    // blueprintPipelines.filter(p => p.status !== TaskStatus.RUNNING).map
     setHistoricalRuns(
       blueprintPipelines.map((p, pIdx) => ({
         id: p.id,
@@ -694,7 +328,6 @@ const BlueprintDetail = (props) => {
         stage: `Stage ${lastPipeline.stage}`,
         tasksFinished: Number(lastPipeline.finishedTasks),
         tasksTotal: Number(lastPipeline.totalTasks),
-        // totalTasks: Number(lastPipeline.totalTasks),
         error: lastPipeline.message || null,
       }))
     }
@@ -833,32 +466,6 @@ const BlueprintDetail = (props) => {
                 </div>
               </div>
             </div>
-
-            {/* <div className='blueprint-connections' style={{ width: '100%', alignSelf: 'flex-start' }}>
-              <h3>Overview</h3>
-              <Card elevation={Elevation.TWO} style={{ padding: '2px' }}>
-              <table className='bp3-html-table bp3-html-table-bordered connections-overview-table' style={{ width: '100%' }}>
-                <thead>
-                  <tr>
-                    <th style={{ minWidth: '200px' }}>Data Connection</th>
-                    <th style={{ width: '100%' }}>Data Scope</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {blueprintConnections?.map((c, cIdx) => (
-                  <tr key={`connection-row-key-${cIdx}`}>
-                    <td>
-                      {c.name}
-                    </td>
-                    <td>
-                      {c.dataScope}{' '}
-                    </td>
-                  </tr>
-                  ))}
-                </tbody>
-              </table>
-              </Card>
-            </div> */}
 
             <div
               className='blueprint-run'
@@ -1016,7 +623,6 @@ const BlueprintDetail = (props) => {
                   >
                     <div
                       className='pipeline-task-activity' style={{
-                      // padding: '20px',
                         flex: 1,
                         padding: Object.keys(currentStages).length === 1 ? '0' : 0,
                         overflow: 'hidden',
@@ -1033,96 +639,7 @@ const BlueprintDetail = (props) => {
                         </div>
                       )}
                     </div>
-                    {/* {currentStages.map((stage, stageIdx) => (
-                      <div
-                        className='run-stage'
-                        key={`run-stage-key-${stageIdx}`}
-                        style={{ flex: 1, margin: '0 4px' }}
-                      >
-                        <h3
-                          className={`stage-header ${stage?.stageHeaderClassName}`}
-                          style={{ margin: '0', padding: '7px' }}
-                        >
-                          <span style={{ float: 'right' }}>{stage?.icon}</span>
-                          {stage?.title}
-                        </h3>
-                        {showCurrentRunTasks && (
-                          <div className='task-activity'>
-                            {stage.tasks.map((stageTask, stIdx) => (
-                              <div
-                                className='stage-task'
-                                key={`stage-task-key-${stIdx}`}
-                                style={{
-                                  display: 'flex',
-                                  flexDirection: 'column',
-                                }}
-                              >
-                                <div
-                                  className='stage-task-info'
-                                  style={{ display: 'flex', padding: '8px' }}
-                                >
-                                  <div
-                                    className='task-icon'
-                                    style={{ minWidth: '24px' }}
-                                  >
-                                    {stageTask.icon}
-                                  </div>
-                                  <div
-                                    className='task-title'
-                                    style={{ flex: 1 }}
-                                  >
-                                    <div style={{ marginBottom: '8px' }}>
-                                      <strong>{stageTask.title}</strong>{' '}
-                                      {stageTask?.caption}
-                                    </div>
-                                    <div
-                                      className='stage-task-progress'
-                                      style={{
-                                        color:
-                                          stageTask?.status ===
-                                          TaskStatus.FAILED
-                                            ? StatusColors.FAILED
-                                            : 'inherit',
-                                      }}
-                                    >
-                                      <div>{stageTask?.message}</div>
-                                      <div>
-                                        {stageTask?.recordsFinished} records
-                                        finished
-                                      </div>
-                                    </div>
-                                  </div>
-                                  <div
-                                    className='task-duration'
-                                    style={{
-                                      display: 'flex',
-                                      justifyContent: 'center',
-                                      alignItems: 'center',
-                                      color: StatusColors[stageTask?.status],
-                                    }}
-                                  >
-                                    {stageTask.duration}{' '}
-                                    {stageTask?.status ===
-                                      TaskStatus.FAILED && (
-                                      <>
-                                        ({TaskStatusLabels[TaskStatus.FAILED]})
-                                      </>
-                                    )}
-                                    {stageTask?.status ===
-                                      TaskStatus.ACTIVE && (
-                                      <>
-                                        ({TaskStatusLabels[TaskStatus.ACTIVE]})
-                                      </>
-                                    )}
-                                  </div>
-                                </div>
-                                <Divider />
-                              </div>
-                            ))}
-                          </div>
-                        )}
-                      </div>
-                    ))} */}
+
                     <Button
                       icon={
                         showCurrentRunTasks ? 'chevron-down' : 'chevron-right'
@@ -1242,20 +759,20 @@ const BlueprintDetail = (props) => {
                               onClick={() => inspectRun(blueprintPipelines.find(p => p.id === run.id))}
                             />
                           </Tooltip>
-                          {/* <Tooltip
+                          <Tooltip
                             intent={Intent.PRIMARY}
                             content='View Full Log'
-                          > */}
-                          <Button
-                            intent={Intent.NONE}
-                            minimal
-                            small
-                            icon='document'
-                            style={{ marginLeft: '10px' }}
-                            // @todo: enable log view dialog support feature
-                            disabled
-                          />
-                          {/* </Tooltip> */}
+                          >
+                            <Button
+                              intent={Intent.NONE}
+                              loading={isDownloading}
+                              minimal
+                              small
+                              icon='document'
+                              style={{ marginLeft: '10px' }}
+                              onClick={() => downloadPipelineLog(blueprintPipelines.find(p => p.id === run.id))}
+                            />
+                          </Tooltip>
                           <Tooltip
                             intent={Intent.PRIMARY}
                             content='Show Run Activity'
